pyproject.toml
[project]
name = "competitionops-agent"
version = "0.1.0"
description = "Google Workspace-native competition PM automation agent for startup teams"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "mcp>=1.0.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/competitionops", "src/competitionops_mcp"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "coverage>=7.6.0"
]
google = [
    "google-api-python-client>=2.140.0",
    "google-auth>=2.33.0",
    "google-auth-oauthlib>=1.2.1"
]
langgraph = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0"
]
otel = [
    "opentelemetry-exporter-otlp>=1.27.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
